Added overload for Throws method in Return and Void behaviours.
Adds an overload for the Throws method that lets the mock's setup code
pass the exception instance that should be thrown upon mock excersice.

This is necessary in cases where the method under test takes some actions
based on the exception properties, enabling more scenarios to be tested.

The main change is in the way the exception is activated, that was
previously placed in the MockProxyDispatcher.Invoke method. To avoid
polluting the CallInfo class, the ExceptionType property was replaced by
the Exception property, which holds the exception to be thrown, whether if
it was set by Throws<T> or Throws(Exception) method.